requirements.txt
openai==0.27.8
Pillow==10.2.0
python-dotenv==1.0.1
Requests==2.31.0
spidev==3.5
